+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Bug fixes
 - **date-picker**: Fix prefilled date-picker display [mpellerin42]
+- **Radio documentation page**: Fix disabled radio group [mpellerin42]
 
 # 2.61.5 (2023-05-15)